Move tables to database.py
