+++ src/flask_server2/app.py
@@ -13,55 +13,20 @@ import glob, os
 import json
 import threading
 
-
-#  - - - - - - - Defaults - - - - - - - # 
-""" 
-Machine specific settings
-    By default the machine should load from the settings to determine the machine state. 
-    The inputs are only to change if needed on startup. 
-    It should check if ./temp/settings.json exists
-    If not it should create it with default settings. 
-    This should not be sent to gitlab, only stored locally 
-"""
-
-app = Flask(__name__)
-
-#  - - - - - - - Database Config and Initialization- - - - - - - # 
-app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql+pymysql://root@127.0.0.1:3306/circuitprinter'
-db = SQLAlchemy(app)
-
-#Base = declarative_base()
-
-
-class Project(db.Model):
-    """
-    A project contains:
-        (string) name, 
-        (list) sequential instances of Process 
-    """
-    #__tablename__ = 'projects'
-    name = db.Column(db.String(100))
-    id = db.Column(db.Integer, primary_key=True)
-    content = db.Column(db.String(300), nullable=False)
-    #slug = db.Column(db.String(100)) 
-
-    # def __init__(self, name):
-    #     self.name = name
-    #     self.slug = '-'.join(name.split()).lower()
-    
-    def __rep__(self):
-        return '<Project %r>' % self.name
+from database import app, db, Project
 
 
 @app.route("/")
 def home():
     return render_template('home.html')
 
+
 def startup():
     pass
 
+
 # = = = = = = = = Projects CRUD = = = = = = = = # 
-@app.route("/projects", methods=["GET","POST"])
+@app.route("/projects", methods=["GET", "POST"])
 def projects():
     # = = = = = = = Create = = = = = = = #
     if request.method == "POST":
@@ -76,13 +41,14 @@ def projects():
             return 'There was an error while adding the project' 
     else:
         # = = = = = = = Read = = = = = = = #
-        projects = Project.query.all()
-        return render_template("projects.html", projects=projects)
+        all_projects = Project.query.all()
+        return render_template("projects.html", projects=all_projects)
 
-@app.route('/delete/<int:id>')
-def delete(id):
+
+@app.route('/delete/<int:project_id>')
+def delete(project_id):
     # = = = = = = = Delete = = = = = = = #
-    project_to_delete = Project.query.get_or_404(id)
+    project_to_delete = Project.query.get_or_404(project_id)
     try:
         db.session.delete(project_to_delete)
         db.session.commit()
@@ -90,10 +56,11 @@ def delete(id):
     except:
         return 'There was an error while deleting that project'
 
-@app.route('/update/<int:id>', methods=['GET','POST'])
-def update(id):
+
+@app.route('/update/<int:project_id>', methods=['GET', 'POST'])
+def update(project_id):
     # = = = = = = = Update = = = = = = = #
-    project = Project.query.get_or_404(id)
+    project = Project.query.get_or_404(project_id)
 
     if request.method == 'POST':
         project.content = request.form['content']
@@ -143,6 +110,7 @@ def settings():
         pass
     else:
         return render_template('settings.html') 
+
 
 @app.route("/calibrate", methods={"GET", "POST"}) 
 def calibrate():
