+++ src/flask_server2/refreshdb.py
@@ -1,3 +1,3 @@
-from app import db  
+from database import db
 
 db.create_all()
